Modify WebSocket to use io.netty library and implement unit testing
Modify WebSocket to use io.netty library and implement unit testing
WebSocket client implementation uses Observer pattern desgin in
order to be able to process every new message received by the server.

Unit testing was implemented checking that the websocket client is
able to subscribe to the three available channels and is able to
retrieve information for each channel.

Signed-off-by: Nelson Victor Cruz Hdez <victor.cruz.isc@gmail.com>